Fix bug where title wasn't passed. Add font.
Title wasn't being passed in as a parameter from the workflow layer to
the activity layer and the activity layer to the service layer. Added a
directory containing Monserrat as a font to use during the video edit.
Explicitly pass in that font into the filter for the overlay. Update
drawtext options by moving title to the top of the screen and using a
bolded type-weight with a border.